feat: added localStorage persistence for Summary and Transcription + created a custom hook for the same
